GPUAggregator : move shader code into seperate files
