+++ modules/core/src/experimental/utils/gpu-grid-aggregation/gpu-grid-aggregator.js
@@ -7,169 +7,12 @@ import {worldToPixels} from 'viewport-mercator-project';
 const {fp64ifyMatrix4} = fp64Utils;
 const IDENTITY_MATRIX = [1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1];
 const PIXEL_SIZE = 4; // RGBA32F
-const AGGREGATE_TO_GRID_VS = `\
-attribute vec2 positions;
-attribute vec2 positions64xyLow;
-attribute float weights;
-uniform vec2 windowSize;
-uniform vec2 cellSize;
-uniform vec2 gridSize;
-uniform mat4 uProjectionMatrix;
-uniform bool projectPoints;
 
-varying float vWeights;
-
-vec2 project_to_pixel(vec2 pos) {
-  vec4 position = vec4(pos, 0., 1.);
-  vec4 result =  uProjectionMatrix * position;
-  return result.xy/result.w;
-}
-
-void main(void) {
-
-  vWeights = weights;
-
-  vec2 windowPos = positions;
-  vec2 windowPos64xyLow = positions64xyLow;
-  if (projectPoints) {
-    windowPos = project_position(windowPos);
-  }
-
-  windowPos = project_to_pixel(windowPos);
-
-  // Transform (0,0):windowSize -> (0, 0): gridSize
-  vec2 pos = floor(windowPos / cellSize);
-
-  // Transform (0,0):gridSize -> (-1, -1):(1,1)
-  pos = (pos * (2., 2.) / (gridSize)) - (1., 1.);
-
-  // Move to pixel center, pixel-size in screen sapce (2/gridSize) * 0.5 => 1/gridSize
-  vec2 offset = 1.0 / gridSize;
-  pos = pos + offset;
-
-  gl_Position = vec4(pos, 0.0, 1.0);
-}
-`;
-
-const AGGREGATE_TO_GRID_VS_FP64 = `\
-attribute vec2 positions;
-attribute vec2 positions64xyLow;
-attribute float weights;
-uniform vec2 windowSize;
-uniform vec2 cellSize;
-uniform vec2 gridSize;
-uniform vec2 uProjectionMatrixFP64[16];
-uniform bool projectPoints;
-
-varying float vWeights;
-
-void project_to_pixel(vec2 pos, vec2 pos64xyLow, out vec2 pixelXY64[2]) {
-
-  vec2 result64[4];
-  vec2 position64[4];
-  position64[0] = vec2(pos.x, pos64xyLow.x);
-  position64[1] = vec2(pos.y, pos64xyLow.y);
-  position64[2] = vec2(0., 0.);
-  position64[3] = vec2(1., 0.);
-  mat4_vec4_mul_fp64(uProjectionMatrixFP64, position64,
-  result64);
-
-  pixelXY64[0] = div_fp64(result64[0], result64[3]);
-  pixelXY64[1] = div_fp64(result64[1], result64[3]);
-}
-
-void main(void) {
-
-  vWeights = weights;
-
-  vec2 windowPos = positions;
-  vec2 windowPos64xyLow = positions64xyLow;
-  if (projectPoints) {
-    vec2 projectedXY[2];
-    project_position_fp64(windowPos, windowPos64xyLow, projectedXY);
-    windowPos.x = projectedXY[0].x;
-    windowPos.y = projectedXY[1].x;
-    windowPos64xyLow.x = projectedXY[0].y;
-    windowPos64xyLow.y = projectedXY[1].y;
-  }
-
-  vec2 pixelXY64[2];
-  project_to_pixel(windowPos, windowPos64xyLow, pixelXY64);
-
-  // Transform (0,0):windowSize -> (0, 0): gridSize
-  vec2 gridXY64[2];
-  gridXY64[0] = div_fp64(pixelXY64[0], vec2(cellSize.x, 0));
-  gridXY64[1] = div_fp64(pixelXY64[1], vec2(cellSize.y, 0));
-  float x = floor(gridXY64[0].x);
-  float y = floor(gridXY64[1].x);
-  vec2 pos = vec2(x, y);
-
-  // Transform (0,0):gridSize -> (-1, -1):(1,1)
-  pos = (pos * (2., 2.) / (gridSize)) - (1., 1.);
-
-  // Move to pixel center, pixel-size in screen sapce (2/gridSize) * 0.5 => 1/gridSize
-  vec2 offset = 1.0 / gridSize;
-  pos = pos + offset;
-
-  gl_Position = vec4(pos, 0.0, 1.0);
-}
-`;
-
-const AGGREGATE_TO_GRID_FS = `\
-precision highp float;
-
-varying float vWeights;
-
-void main(void) {
-  gl_FragColor = vec4(1., vWeights, 0, 0.0);
-}
-`;
-
-const AGGREGATE_ALL_VS_FP64 = `\
-#version 300 es
-
-in vec2 position;
-uniform vec2 gridSize;
-
-out vec2 vTextureCoord;
-void main(void) {
-  // Map each position to single pixel
-  vec2 pos = vec2(-1.0, -1.0);
-
-  // Move to pixel center, pixel-size in screen sapce (2/gridSize) * 0.5 => 1/gridSize
-  vec2 offset = 1.0 / gridSize;
-  pos = pos + offset;
-
-  gl_Position = vec4(pos, 0.0, 1.0);
-
-  float yIndex = floor(float(gl_InstanceID) / gridSize[0]);
-  float xIndex = float(gl_InstanceID) - (yIndex * gridSize[0]);
-
-  vec2 yIndexFP64 = vec2(yIndex, 0.);
-  vec2 xIndexFP64 = vec2(xIndex, 0.);
-  vec2 gridSizeYFP64 = vec2(gridSize[1], 0.);
-  vec2 gridSizeXFP64 = vec2(gridSize[0], 0.);
-
-  vec2 texCoordXFP64 = div_fp64(yIndexFP64, gridSizeYFP64);
-  vec2 texCoordYFP64 = div_fp64(xIndexFP64, gridSizeXFP64);
-
-  vTextureCoord = vec2(texCoordYFP64.x, texCoordXFP64.x);
-}
-`;
-
-const AGGREGATE_ALL_FS = `\
-#version 300 es
-precision highp float;
-
-in vec2 vTextureCoord;
-uniform sampler2D uSampler;
-out vec4 fragColor;
-void main(void) {
-  vec4 textureColor = texture(uSampler, vec2(vTextureCoord.s, vTextureCoord.t));
-  // Red: total count, Green: total weight, Alpha: maximum wieght
-  fragColor = vec4(textureColor.r, textureColor.g, 0., textureColor.g);
-}
-`;
+import AGGREGATE_TO_GRID_VS from './aggregate-to-grid-vs.glsl';
+import AGGREGATE_TO_GRID_VS_FP64 from './aggregate-to-grid-vs-64.glsl';
+import AGGREGATE_TO_GRID_FS from './aggregate-to-grid-fs.glsl';
+import AGGREGATE_ALL_VS_FP64 from './aggregate-all-vs-64.glsl';
+import AGGREGATE_ALL_FS from './aggregate-all-fs.glsl';
 
 const DEFAULT_CHANGE_FLAGS = {
   dataChanged: true,
